+++ common/src/filterFunctions.c
@@ -36,19 +36,19 @@ void	init_filter (void)
 	
 }
 
-short highpass_filter (short next_sample)
+float highpass_filter (float next_sample)
 { 
 	xv[0] = xv[1]; xv[1] = xv[2]; 
-	xv[2] = (float) next_sample;
+	xv[2] = next_sample;
 	yv[0] = yv[1]; yv[1] = yv[2]; 
 	
 	yv[2] = (xv[0] + xv[2]) - xv[1] - xv[1]
 		+ ( -0.8752145483 * yv[0]) + (  1.8668922797 * yv[1]);
   
-	return (short) yv[2];
+	return yv[2];
 }
 
-short notch_filter_100 (short next_sample)
+float notch_filter_100 (float next_sample)
 { 
 	xv[0] = xv[1]; xv[1] = xv[2]; xv[2] = xv[3]; xv[3] = xv[4]; 
   xv[4] = next_sample;
@@ -57,10 +57,10 @@ short notch_filter_100 (short next_sample)
                      + ( -0.9858833621 * yv[0]) + (  3.8332644592 * yv[1])
                      + ( -5.7119092296 * yv[2]) + (  3.8606110802 * yv[3]);
         
-	return (short) yv[4];
+	return yv[4];
 }
 
-short notch_filter_120 (short next_sample)
+float notch_filter_120 (float next_sample)
 { 
 	xv[0] = xv[1]; xv[1] = xv[2]; xv[2] = xv[3]; xv[3] = xv[4]; 
   xv[4] = next_sample;
@@ -69,6 +69,6 @@ short notch_filter_120 (short next_sample)
                      + ( -0.9858833621 * yv[0]) + (  3.7789716667 * yv[1])
                      + ( -5.6071060574 * yv[2]) + (  3.8059309614 * yv[3]);
         
-	return (short) yv[4];
+	return yv[4];
 }
